player attack , coin gathering, Player con

+++ core/src/com/rpg/game/entities/Coin.java
@@ -9,7 +9,7 @@ public class Coin extends Entity {
 	private String bodyTAG = "coin";
 	private boolean isSensor = true;
 	
-	public Coin(int x, int y) {
+	public Coin(float x, float y) {
 	
 		enemycreator(x*PPM, y*PPM,bodyTAG,B2DVars.BIT_ENEMY, B2DVars.BIT_ENEMY| B2DVars.BIT_PLAYER,  isSensor );
 		playAnimation(getAnimationRow(),textureName);
